Split ApolloFeatureDetails into multiple components
