+++ packages/jbrowse-plugin-apollo/src/FeatureDetailsWidget/model.ts
@@ -1,14 +1,21 @@
 import { ElementId } from '@jbrowse/core/util/types/mst'
 import { AnnotationFeature } from 'apollo-mst'
-import { types } from 'mobx-state-tree'
+import { Instance, SnapshotIn, types } from 'mobx-state-tree'
 
-import { ChangeManager } from '../ChangeManager'
+export const ApolloFeatureDetailsWidgetModel = types.model(
+  'ApolloFeatureDetailsWidget',
+  {
+    id: ElementId,
+    type: types.literal('ApolloFeatureDetailsWidget'),
+    feature: types.safeReference(AnnotationFeature),
+    assembly: types.string,
+    refName: types.string,
+  },
+)
 
-export const ApolloFeatureDetails = types.model('ApolloFeatureDetails', {
-  id: ElementId,
-  type: types.literal('ApolloFeatureDetails'),
-  feature: types.safeReference(AnnotationFeature),
-  assembly: types.string,
-  refName: types.string,
-  changeManager: types.frozen<ChangeManager>(),
-})
+export type ApolloFeatureDetailsWidget = Instance<
+  typeof ApolloFeatureDetailsWidgetModel
+>
+export type ApolloFeatureDetailsWidgetSnapshot = SnapshotIn<
+  typeof ApolloFeatureDetailsWidgetModel
+>
